fix variable type to avoid wrong lengths in the untangled PAF

+++ src/algorithms/untangle.cpp
@@ -937,7 +937,7 @@ void untangle(
 #pragma omp parallel for schedule(dynamic, 1) num_threads(num_threads)
         for (uint64_t i = 0; i < paths.size(); ++i) {
             auto& path = paths[i];
-            const uint16_t path_len = get_path_length(graph, paths[i]);
+            const uint64_t path_len = get_path_length(graph, paths[i]);
 
             // You can't write on such a data structure in parallel
 #pragma omp critical (path_to_len)
